stele

Enum MetricValue

pub enum MetricValue {
    Count {
        value: f64,
    },
    Rate {
        interval: u64,
        value: f64,
    },
    Gauge {
        value: f64,
    },
    Sketch {
        sketch: DDSketch,
    },
}

A metric value.

Equality

MetricValue implements PartialEq and Eq, the majority of which involves comparing floating-point (f64) numbers. Comparing floating-point numbers for equality is inherently tricky (this is just one blog post/article out of thousands on the subject). In the equality implementation for MetricValue, we use a ratio-based approach.

This means that when comparing two floating-point numbers, we look at their ratio to one another, with an upper bound on the allowed difference. For example, if we compare 99 to 100, thereโs a difference of 1% (1 - (99/100) = 0.01 = 1%), while the difference between 99.999 and 100 is only 0.001% (1 - (99.999/100) = 0.00001 = 0.001%). As most comparisons are expected to be close, only differing by a few ULPs (units in the last place) due to slight differences in how floating-point numbers are implemented between Go and Rust, this approach is sufficient to compensate for the inherent imprecision while not falling victim to relying on ULPs or epsilon directly, whose applicability depends on the number range being compared.

Specifically, we compare floating-point numbers using a ratio of 0.00000001 (0.0000001%), meaning the smaller of the two values being compared must be within 99.999999% to 100% of the larger number, which is sufficiently precise for our concerns.

Variants

Count

A count.

Fields

value: f64

The value of the count.

Rate

A rate.

Rates are per-second adjusted counts. For example, a count that increased by 100 over 10 seconds would be represented as a rate with an interval of 10 (seconds) and a value of 10 (100 / 10 = 10).

Fields

interval: u64

The interval of the rate, in seconds.

value: f64

The per-second value of the rate.

Gauge

A gauge.

Fields

value: f64

The value of the gauge.

Sketch

A sketch.

Fields

sketch: DDSketch

The sketch data.